td3: add some trained models for quick testing
The saved model is that of the actor network. The naming follows this
convention:

    td3_<checkpoint>_<episode>_actor.pth

<checkpoint>==1 is for 1d actor and <checkpoint>==2 is for 2d actor. NOTE
that the numbers being the same as the action-space dimension is
unrelated. I.e., <checkpoint>==3 won't necessarily mean the action space
is 3d. <checkpoint> is just that, a checkpoint, where we have achieved a
significant amount of progress to warrant saving the models.

Use the model with the highest <episode>. The model is saved only when
the total episodic return is greater than the max of the episodic
returns of all previous episodes. Which means, the model with the
highest <episode> number yields the greatest total episodic return.